readme-requirements.txt
default admin account
Username: 123
Password: 123

Backup admin (just incase, password is weird because of the password specs)
Username: admin
Passowrd: adm1n-141


------ Requirements -----

djangorestframework
django
